+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ## Unreleased
 
+- Convert h1 headers to h2, convert h4 and h5 headers to h3 (PR #41)
 - Fix dist file containing ES2015 and regex breaking in IE11 (PR #39)
 
 ## 0.1.1