+++ CHANGELOG.md
@@ -8,7 +8,7 @@
   useful summary for people upgrading their application, not a replication
   of the commit log.
 
-## Unreleased
+## 0.2.0
 
 - Convert Microsoft Word lists to markdown lists (PR #42)
 - Convert h1 headers to h2, convert h4 and h5 headers to h3 (PR #41)